backend/requirements.txt
fastapi[all]
uvicorn
python-multipart
Pillow==9.4.0
matplotlib==3.6.2
numpy==1.23.4
opencv-python
tensorflow==2.10.1